Clean up, small fixes and improvements
now the lanescore resets correctly and the game ends when all lanes fail. The gameboard now has the variable to control number of lanes and I tried to clean up the files a bit. The lanes now show what key to press. There are issues with colors not looking great and some things not being alligned.